fix: moved variable for ETHEREUM_EXTENSION
