+++ src/components/assets/Account.vue
@@ -154,7 +154,6 @@ import AuIcon from 'src/components/header/modals/account-unification/AuIcon.vue'
 import { endpointKey, providerEndpoints } from 'src/config/chainEndpoints';
 import { SupportWallet, supportWalletObj } from 'src/config/wallets';
 import {
-  ETHEREUM_EXTENSION,
   useAccount,
   useAccountUnification,
   useBalance,
@@ -162,6 +161,7 @@ import {
   useNetworkInfo,
   useWalletIcon,
 } from 'src/hooks';
+import { ETHEREUM_EXTENSION } from 'src/modules/account';
 import { useEvmAccount } from 'src/hooks/custom-signature/useEvmAccount';
 import { getEvmMappedSs58Address, setAddressMapping } from 'src/hooks/helper/addressUtils';
 import { useDappStaking } from 'src/staking-v3';
